# Simula/simul/simulphase2backend/docs/swagger.json
{ "openapi":"3.0.0",
  "info":{
    "title":"Simul Collaborative Design Platform Database",
    "description":"This page displays some of the key routes of the API, and shows what is returned",
    "version":"1.0.0"
  },
  "basePath":"/api",
...
  "paths":{
    "/register":{
      "post":{
        "tags":["User Authentication"],
        "description":"This is where users can create an account.",
        "operationId":"register",
        "produces":["application/json"],
        "requestBody":{
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name":{
                    "type": "string",
                    "example": "Steve Stephenson",
                    "description":"Name",
                    "required":true
                  },
                  "phone":{
                    "type": "string",
                    "example": "04111111111",
                    "description":"Phone",
                    "required":true
                  },
                  "email":{
                    "type": "string",
                    "example": "demo-user@gmail.com",
                    "description":"Email",
                    "required":true
                  },
                  "password":{
                    "type": "string",
                    "example": "demouserpassword",
                    "description":"Password for authentication",
                    "required":true
                  }
                }
              }
            }
          }
        },
        "responses":{
          "200": {
              "description": "Created",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "string",
                        "example": "success"
                      },
                      "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiZGVtby11c2VyQGdtYWlsLmNvbWZnQkJiU0lSa3Y3eSs2dndUeWhPdDA5c294UEpZYy9sNG1LR1d3RlY2WGc9IiwiaWF0IjoxNTg0NDEyMTMyLCJleHAiOjE1ODQ0OTg1MzJ9.x-718Nsldfs6BAJP5sNng_uiMuX3g9T8bYQYxgSHdW0"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Error. Name is required"
                      }
                    }
                  }
                }
              }
            }
          }
        }
    },
...
    "/createproject":{
      "post":{
        "tags":["Projects"],
        "description":"Creating a project with client credentials",
        "operationId":"createProject",
        "produces":["application/json"],
        "requestBody":{
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name":{
                    "type": "string",
                    "example": "Ben Smith",
                    "description":"Client Name",
                    "required":true
                  },
                  "phone":{
                    "type": "string",
                    "example": "04001 234 567",
                    "description":"Client Phone",
                    "required":true
                  },
                  "email":{
                    "type": "string",
                    "example": "bensmith@gmail.com",
                    "description":"Client Email",
                    "required":true
                  },
                  "projectAddress":{
                    "type": "string",
                    "example": "6 ProjectCreated Avenue",
                    "description":"Address of Project",
                    "required":true
                  },
                  "projectedCompletionDate": {
                    "type": "string",
                    "example": "2020/04/12",
                    "description":"Projected Completion Date for Project",
                    "required":true
                  },
                  "userID": {
                    "type": "number",
                    "example": 1,
                    "description":"The ID of the user logged in",
                    "required":true
                  }
                }
              }
            }
          }
        },
        "responses":{
          "200": {
              "description": "Created",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "You have successfully created a project."
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "oops! It looks like the project can't be created :("
                      }
                    }
                  }
                }
              }
            }
          }
        }
    },
...
    }
  }